Move removeMsgDups to main utils
