+++ src/utils/apollo/defaultCacheConfig.ts
@@ -1,4 +1,5 @@
 import { InMemoryCacheConfig } from '@apollo/client'
+import { removeMessageDups } from '../removeGQLMsgDups'
 
 // The params field is expected to be a JSON string
 // or null. Both are safe to pass to JSON parse. The
@@ -30,26 +31,6 @@ const parseExecutionTrace = (_: any, incoming: any) => {
     console.error(e)
     return null
   }
-}
-
-export function removeMessageDups(
-  existing: { __ref: any }[],
-  incoming: { __ref: any }[],
-  args: { offset: number; limit: number }
-) {
-  const head = [...existing]
-  const tail = head.splice(args.offset * args.limit)
-  const merged = [].concat(head, incoming, tail)
-
-  const found = new Set([])
-  return [...merged].filter(ele => {
-    if (!found.has(ele.__ref)) {
-      found.add(ele.__ref)
-      return true
-    }
-
-    return false
-  })
 }
 
 export const defaultMessageHistoryClientCacheConfig: InMemoryCacheConfig = {
